backend/flask/requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
uWSGI==2.0.18
Werkzeug==0.15.2
